Split the types used to build MIR values from the Transformer logic that traverses the AST
